updated package validation to allow subfolders

+++ data_management/services/dlu_filesystem.py
@@ -65,24 +65,11 @@ class DirectoryInfo:
                 self.subdir_count += 1
             else:
                 self.file_count += 1
-                checksum = "0" if self.calculate_checksums else calculate_checksum(full_path)
-                self.file_details.append(DLUFile(item, full_path, checksum, os.path.getsize(full_path)))
+            checksum = "0" if self.calculate_checksums else calculate_checksum(full_path)
+            self.file_details.append(DLUFile(item, full_path, checksum, os.path.getsize(full_path)))
 
     def check_if_valid_for_dlu(self):
-        directory_not_empty = len(self.dir_contents) != 0
-        # Nothing but a subdir
-        if self.subdir_count == 1 and self.file_count == 0:
-            subdir_path = os.path.join(self.directory_path, self.dir_contents[0])
-            subdir_info = DirectoryInfo(subdir_path, self.calculate_checksums)
-            if subdir_info.subdir_count == 0 and subdir_info.valid_for_dlu:
-                self.valid_for_dlu = True
-            else:
-                self.valid_for_dlu = False
-        # Items and no subdirectories
-        elif self.subdir_count == 0 and directory_not_empty:
-            self.valid_for_dlu = True
-        else:
-            self.valid_for_dlu = False
+        return len(self.dir_contents) != 0
 
 
 class DLUFileHandler:
@@ -105,9 +92,13 @@ class DLUFileHandler:
                 os.makedirs(dest_package_directory, exist_ok=True)
             source_file = os.path.join(source_package_directory, file.name)
             dest_file = os.path.join(dest_package_directory, file.name)
-            logger.info("Copying file to " + dest_file)
             if not os.path.exists(dest_file):
-                shutil.copy(source_file, dest_file)
+                if os.path.isdir(source_file):
+                    logger.info("Copying directory to " + dest_file)
+                    shutil.copytree(source_file, dest_file)
+                elif os.path.isfile(source_file):
+                    logger.info("Copying file to " + dest_file)
+                    shutil.copy(source_file, dest_file)
                 files_copied = files_copied + 1
             else:
                 logger.warning(dest_file + " already exists. Skipping.")
@@ -117,19 +108,10 @@ class DLUFileHandler:
         logger.info("Moving files for package " + package_id)
         source_package_directory = self.globus_data_directory + '/' + package_id
         source_directory_info = DirectoryInfo(source_package_directory, False)
-        success = False
+        success = True
 
-        # Make sure the directory is not empty and does not have more than one subdirectory.
-        if source_directory_info.valid_for_dlu:
-            # Set the source path to the subdirectory if it has only one and is valid.
-            if source_directory_info.subdir_count == 1 and source_directory_info.file_count == 0:
-                source_package_directory = os.path.join(source_package_directory,
-                                                        source_directory_info.dir_contents[0])
-                source_directory_info = DirectoryInfo(source_package_directory, False)
-                logger.info(
-                    "Found one subdirectory (" + source_package_directory + "). Setting it as the main data directory.")
-            success = True
-        else:
+        # Make sure the directory is not empty
+        if not source_directory_info.valid_for_dlu:
             success = False
             logger.error("Directory for package " + package_id + " failed validation.")
         return success
